+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+### v0.0.38
+
+- Add Core Milestone Component
+- Added caret icon for picker on android
+
 ### v0.0.37
 
-- Add FilterDrawer component
+- Modify the Data block component adding JSX element for left and right.
+- Added Circular progress view
 - Add Right Element and Expandable element in Compact-card-list
 - Fix Avatar component prop heirarchy
 - Fix PeopleTabs spacings